move data gen function
